requirements.txt
flask
tensorflow
keras
pillow
gradio
streamlit